refactor(listen): split explore views
